Remove unused code (from Fleet's sandbox implementation)
