+++ cmd/fleet/serve.go
@@ -189,7 +189,6 @@ the way that the Fleet server works.
 
 			var ds fleet.Datastore
 			var carveStore fleet.CarveStore
-			var installerStore fleet.InstallerStore
 
 			opts := []mysql.DBOption{mysql.Logger(logger), mysql.WithFleetConfig(&config)}
 			if config.MysqlReadReplica.Address != "" {
@@ -219,14 +218,6 @@ the way that the Fleet server works.
 				}
 			} else {
 				carveStore = ds
-			}
-
-			if config.Packaging.S3.Bucket != "" {
-				var err error
-				installerStore, err = s3.NewInstallerStore(config.Packaging.S3)
-				if err != nil {
-					initFatal(err, "initializing S3 installer store")
-				}
 			}
 
 			migrationStatus, err := ds.MigrationStatus(cmd.Context())
@@ -733,7 +724,6 @@ the way that the Fleet server works.
 				ssoSessionStore,
 				liveQueryStore,
 				carveStore,
-				installerStore,
 				failingPolicySet,
 				geoIP,
 				redisWrapperDS,
