+++ server/datastore/s3/s3.go
@@ -12,6 +12,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/fleetdm/fleet/v4/server/config"
+	"github.com/fleetdm/fleet/v4/server/fleet"
 )
 
 const awsRegionHint = "us-east-1"
@@ -21,6 +22,18 @@ type s3store struct {
 	bucket           string
 	prefix           string
 	cloudFrontConfig *config.S3CloudFrontConfig
+}
+
+type installerNotFoundError struct{}
+
+var _ fleet.NotFoundError = (*installerNotFoundError)(nil)
+
+func (p installerNotFoundError) Error() string {
+	return "installer not found"
+}
+
+func (p installerNotFoundError) IsNotFound() bool {
+	return true
 }
 
 // newS3store initializes an S3 Datastore
