+++ server/datastore/s3/testing_utils.go
@@ -1,24 +1,19 @@
 package s3
 
 import (
-	"context"
 	"os"
-	"strings"
 	"testing"
 
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/fleetdm/fleet/v4/server/config"
-	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/stretchr/testify/require"
 )
 
 const (
-	accessKeyID           = "minio"
-	secretAccessKey       = "minio123!"
-	testEndpoint          = "localhost:9000"
-	mockInstallerContents = "mock"
+	accessKeyID     = "minio"
+	secretAccessKey = "minio123!"
+	testEndpoint    = "localhost:9000"
 )
 
 func SetupTestSoftwareInstallerStore(tb testing.TB, bucket, prefix string) *SoftwareInstallerStore {
@@ -29,14 +24,6 @@ func SetupTestSoftwareInstallerStore(tb testing.TB, bucket, prefix string) *Soft
 
 func SetupTestBootstrapPackageStore(tb testing.TB, bucket, prefix string) *BootstrapPackageStore {
 	store := setupTestStore(tb, bucket, prefix, NewBootstrapPackageStore)
-	tb.Cleanup(func() { cleanupStore(tb, store.s3store) })
-	return store
-}
-
-// SetupTestInstallerStore creates a new store with minio as a back-end
-// for local testing
-func SetupTestInstallerStore(tb testing.TB, bucket, prefix string) *InstallerStore {
-	store := setupTestStore(tb, bucket, prefix, NewInstallerStore)
 	tb.Cleanup(func() { cleanupStore(tb, store.s3store) })
 	return store
 }
@@ -73,37 +60,6 @@ func setupTestStore[T testBucketCreator](tb testing.TB, bucket, prefix string, n
 	require.NoError(tb, err)
 
 	return store
-}
-
-// SeedTestInstallerStore adds mock installers to the given store
-func SeedTestInstallerStore(tb testing.TB, store *InstallerStore, enrollSecret string) []fleet.Installer {
-	checkEnv(tb)
-	installers := []fleet.Installer{
-		mockInstaller(enrollSecret, "pkg", true),
-		mockInstaller(enrollSecret, "msi", true),
-		mockInstaller(enrollSecret, "deb", true),
-		mockInstaller(enrollSecret, "rpm", true),
-		mockInstaller(enrollSecret, "pkg", false),
-		mockInstaller(enrollSecret, "msi", false),
-		mockInstaller(enrollSecret, "deb", false),
-		mockInstaller(enrollSecret, "rpm", false),
-	}
-
-	for _, i := range installers {
-		_, err := store.Put(context.Background(), i)
-		require.NoError(tb, err)
-	}
-
-	return installers
-}
-
-func mockInstaller(secret, kind string, desktop bool) fleet.Installer {
-	return fleet.Installer{
-		EnrollSecret: secret,
-		Kind:         kind,
-		Desktop:      desktop,
-		Content:      aws.ReadSeekCloser(strings.NewReader(mockInstallerContents)),
-	}
 }
 
 func cleanupStore(tb testing.TB, store *s3store) {
