+++ server/service/service.go
@@ -30,7 +30,6 @@ type Service struct {
 	ds             fleet.Datastore
 	task           *async.Task
 	carveStore     fleet.CarveStore
-	installerStore fleet.InstallerStore
 	resultStore    fleet.QueryResultStore
 	liveQueryStore fleet.LiveQueryStore
 	logger         kitlog.Logger
@@ -98,7 +97,6 @@ func NewService(
 	sso sso.SessionStore,
 	lq fleet.LiveQueryStore,
 	carveStore fleet.CarveStore,
-	installerStore fleet.InstallerStore,
 	failingPolicySet fleet.FailingPolicySet,
 	geoIP fleet.GeoIP,
 	enrollHostLimiter fleet.EnrollHostLimiter,
@@ -117,7 +115,6 @@ func NewService(
 		ds:                ds,
 		task:              task,
 		carveStore:        carveStore,
-		installerStore:    installerStore,
 		resultStore:       resultStore,
 		liveQueryStore:    lq,
 		logger:            logger,
